Extract opt package
